#2 feat: 검색부분 팝업창 컴포넌트 분리로 관리
